requirements.txt
# Core
#python-dotenv>=1.0.0 
pyyaml==6.0.1
loguru==0.7.2

# Data & ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0

# Visualization
plotly==5.22.0
matplotlib==3.9.0
seaborn==0.13.2

# Web UI
streamlit==1.36.0

# Database
sqlalchemy==2.0.31

#psycopg2-binary==2.9.9
psycopg2-binary>=2.9


# Monitoring
prometheus-client==0.20.0

# Utilities
tqdm==4.66.4
